+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 * Rename `DraftUtils.addLineBreakRemovingSelection` to `DraftUtils.addLineBreak`.
+* Replace `showUndoRedoControls` with separate props `showUndoControl` and `showRedoControl` for which control UI can be overriden.
 
 ### Fixed